Addition of WrapperObject.java and adjusting existing classes to fit
Decided to add a new subclass of RoomObject called WrapperObject, which will be similar to ContainerObject but where ContainerObject represents a chair that has a key on it (or something like that), WrapperObject would be a bar of chocolate that can be acquired. It wraps a single item, and hides itself when that item is taken.